Convert antiques to MUI card grid
I have tried this before, but had hangups in the image sizing.
Thanks for the guidance from Ashutosh Singh of Pluralsight:
https://www.pluralsight.com/guides/styling-a-react-app-with-material-ui

It's not perfect now (grid has no margin, images not centered)
because I need to get better at overriding MUI classes, but this
is a good start.

Addresses #7 (Convert "Unsold Stock" page to MUI grid of cards)
Project: "Port UI to MUI"