Reimplement `Search FC` with hooks
I was too focused on trying to make some approach where all of the
hooks were in `Search FC`. Rather than that, I found a tutorial
where the search was merely presentational logic which called
hooks in the parent. This was similar to class based state handling
that I had done, but now with hooks. This is fine, for now...

Thank you to Emma Goto with the article "How to build a search bar
in React':
https://www.emgoto.com/react-search-bar/